(XKB Bug workaround) Send a fake `Ctrl` after applying the XKB mapping
Check https://gitlab.freedesktop.org/xorg/app/xkbcomp/-/issues/9 for
more info.

I was observing this behavior when trying to use the
`interception_tools` + `space2meta` (adapted to send a RCtrl) only with
the laptop keyboard.

The behavior was that the custom XKB mapping wasn't being applied to the
"new" device that was "plugged" (in this case it was the virtual device
created by the `interception`).